Created dungeon crawler project; renamed monogame library
