+++ Caps.RPG.DungeonCrawler/Caps.RPG.DungeonCrawler.csproj
@@ -23,12 +23,13 @@
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Gum.MonoGame" Version="2025.5.13.1" />
     <PackageReference Include="MonoGame.Framework.DesktopGL" Version="3.8.*" />
     <PackageReference Include="MonoGame.Content.Builder.Task" Version="3.8.*" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Caps.RPG.MonoGame\Caps.RPG.MonoGame.csproj" />
+    <Folder Include="Metadata\" />
+    <Folder Include="UI\" />
+    <Folder Include="Scenes\" />
   </ItemGroup>
   <Target Name="RestoreDotnetTools" BeforeTargets="Restore">
     <Message Text="Restoring dotnet tools" Importance="High" />
