+++ Caps.RPG.DungeonCrawler/Metadata/Attributions.txt
@@ -1,0 +1,1 @@
+Icon - “Designed by Freepik”
