+++ Caps.RPG.DungeonCrawler/Program.cs
@@ -1,0 +1,11 @@
+﻿namespace Caps.RPG.DungeonCrawler
+{
+    internal static class Program
+    {
+        private static void Main()
+        {
+            using var game = new DungeonCrawlerClient();
+            game.Run();
+        }
+    }
+}
