+++ Caps.RPG.GameApp/Scenes/TitleScene.cs
@@ -1,5 +1,8 @@
 ﻿using System;
 using Caps.RPG.GameApp.UI;
+using Caps.RPG.MonoGame;
+using Caps.RPG.MonoGame.Graphics;
+using Caps.RPG.MonoGame.Scenes;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Audio;
 using Microsoft.Xna.Framework.Graphics;
@@ -7,9 +10,6 @@ using Microsoft.Xna.Framework.Input;
 using MonoGameGum;
 using MonoGameGum.Forms.Controls;
 using MonoGameGum.GueDeriving;
-using MonoGameLibrary;
-using MonoGameLibrary.Graphics;
-using MonoGameLibrary.Scenes;
 
 namespace Caps.RPG.GameApp.Scenes
 {
