+++ Caps.RPG.GameApp/UI/OptionsSlider.cs
@@ -1,11 +1,11 @@
 ﻿using System;
+using Caps.RPG.MonoGame.Graphics;
 using Gum.DataTypes;
 using Gum.DataTypes.Variables;
 using Gum.Managers;
 using Microsoft.Xna.Framework;
 using MonoGameGum.Forms.Controls;
 using MonoGameGum.GueDeriving;
-using MonoGameLibrary.Graphics;
 
 namespace Caps.RPG.GameApp.UI
 {
