+++ Caps.RPG.MonoGame/Audio/AudioController.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using Microsoft.Xna.Framework.Audio;
 using Microsoft.Xna.Framework.Media;
 
-namespace MonoGameLibrary.Audio
+namespace Caps.RPG.MonoGame.Audio
 {
     public class AudioController : IDisposable
     {
