+++ Caps.RPG.MonoGame/Graphics/AnimatedSprite.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Microsoft.Xna.Framework;
 
-namespace MonoGameLibrary.Graphics
+namespace Caps.RPG.MonoGame.Graphics
 {
     public class AnimatedSprite : Sprite
     {
