+++ Caps.RPG.MonoGame/Graphics/Animation.cs
@@ -1,10 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace MonoGameLibrary.Graphics
+namespace Caps.RPG.MonoGame.Graphics
 {
     public class Animation
     {
