+++ Caps.RPG.MonoGame/Graphics/TextureRegion.cs
@@ -1,12 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace MonoGameLibrary.Graphics
+namespace Caps.RPG.MonoGame.Graphics
 {
     public class TextureRegion
     {
