+++ Caps.RPG.MonoGame/Input/GamePadInfo.cs
@@ -2,7 +2,7 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Input;
 
-namespace MonoGameLibrary.Input
+namespace Caps.RPG.MonoGame.Input
 {
     public class GamePadInfo
     {
