+++ Caps.RPG.MonoGame/Input/MouseInfo.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Input;
 
-namespace MonoGameLibrary.Input
+namespace Caps.RPG.MonoGame.Input
 {
     public class MouseInfo
     {
