+++ Caps.RPG.sln
@@ -11,7 +11,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Caps.RPG.Content", "Caps.RP
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Caps.RPG.GameApp", "Caps.RPG.GameApp\Caps.RPG.GameApp.csproj", "{43E242D6-BEE8-438B-98D5-F439B0CF818D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MonoGameLibrary", "MonoGameLibrary\MonoGameLibrary.csproj", "{496E8707-AA8E-4E92-92B6-98E2D04E8B31}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Caps.RPG.DungeonCrawler", "Caps.RPG.DungeonCrawler\Caps.RPG.DungeonCrawler.csproj", "{06924811-F9AC-4D5D-8F18-54B711703C52}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Caps.RPG.MonoGame", "Caps.RPG.MonoGame\Caps.RPG.MonoGame.csproj", "{2DEAFE19-8EBA-4A3A-8647-98C34015CEF1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -35,10 +37,14 @@ Global
 		{43E242D6-BEE8-438B-98D5-F439B0CF818D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{43E242D6-BEE8-438B-98D5-F439B0CF818D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{43E242D6-BEE8-438B-98D5-F439B0CF818D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{496E8707-AA8E-4E92-92B6-98E2D04E8B31}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{496E8707-AA8E-4E92-92B6-98E2D04E8B31}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{496E8707-AA8E-4E92-92B6-98E2D04E8B31}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{496E8707-AA8E-4E92-92B6-98E2D04E8B31}.Release|Any CPU.Build.0 = Release|Any CPU
+		{06924811-F9AC-4D5D-8F18-54B711703C52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{06924811-F9AC-4D5D-8F18-54B711703C52}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{06924811-F9AC-4D5D-8F18-54B711703C52}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{06924811-F9AC-4D5D-8F18-54B711703C52}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2DEAFE19-8EBA-4A3A-8647-98C34015CEF1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2DEAFE19-8EBA-4A3A-8647-98C34015CEF1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2DEAFE19-8EBA-4A3A-8647-98C34015CEF1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2DEAFE19-8EBA-4A3A-8647-98C34015CEF1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
